Make linux build functional again. (#18)
* upd: Switch to quill for logging

* upd: Add linux binaries for glfw

* chore: Remove glm static libraries

* chore: Add linux binaries for fmod

* chore: Remove build artifacts from directory

* chore: Add manual build of imgui

* chore: Make linux build and run working again

* chore: Get logging with quill working

* upd: Add missing glfw prebuilts

* fix: Segfault after closing the editor

+++ Source/Runtime/polos/core/ecs/component_memory.cpp
@@ -14,7 +14,7 @@ namespace polos::ecs
 
         std::size_t cumulative_size = 0;
 
-        for (int i = 0; i < MAX_COMPONENT_COUNT_FOR_ENTITY; ++i)
+        for (size_t i = 0; i < MAX_COMPONENT_COUNT_FOR_ENTITY; ++i)
         {
             component_offsets[i] = cumulative_size;
             cumulative_size += component_size[i] * MAX_ENTITY_COUNT_IN_SCENE;
